+++ packages/ag-charts-community/src/chart/label.ts
@@ -11,7 +11,7 @@ import type { Matrix } from '../scene/matrix';
 import type { PlacedLabelDatum } from '../scene/util/labelPlacement';
 import { normalizeAngle360, toRadians } from '../util/angle';
 import { BaseProperties } from '../util/properties';
-import { TextUtils } from '../util/textMeasurer';
+import { type TextMeasurer, TextUtils } from '../util/textMeasurer';
 import type { RequireOptional } from '../util/types';
 import {
     BOOLEAN,
@@ -23,6 +23,7 @@ import {
     STRING,
     Validate,
 } from '../util/validation';
+import type { TickDatum } from './axis/axis';
 import type { ChartAxisLabelFlipFlag } from './chartAxis';
 
 export class Label<TParams = never, TDatum = any>
@@ -128,19 +129,30 @@ export function getTextAlign(
     return 'start';
 }
 
-export function calculateLabelBBox(text: string, bbox: BBox, labelMatrix: Matrix): PlacedLabelDatum {
-    // Text.getBBox() does not take into account any of the transformations that have been applied to the label nodes, only the width and height are useful.
-    // Rather than taking into account all transformations including those of parent nodes which would be the result of `computeTransformedBBox()`, giving the x and y in the entire axis coordinate space,
-    // take into account only the rotation and translation applied to individual label nodes to get the x y coordinates of the labels relative to each other
-    // this makes label collision detection a lot simpler
-    const { x: labelX, y: labelY, width, height } = bbox;
-    const translatedBBox = new BBox(labelX, labelY, 0, 0);
+export function createLabelData(
+    tickData: TickDatum[],
+    labelX: number,
+    labelMatrix: Matrix,
+    textMeasurer: TextMeasurer
+) {
+    const labelData: PlacedLabelDatum[] = [];
 
-    labelMatrix.transformBBox(translatedBBox, bbox);
+    for (const { tickLabel: text, translationY } of tickData) {
+        if (!text) continue;
 
-    const { x, y } = bbox;
-    return {
-        point: { x, y },
-        label: { text, width, height },
-    };
+        const { width, height } = textMeasurer.measureLines(text);
+        const bbox = new BBox(labelX, translationY, width, height);
+        const translatedBBox = new BBox(labelX, translationY, 0, 0);
+
+        labelMatrix.transformBBox(translatedBBox, bbox);
+
+        const { x, y } = bbox;
+
+        labelData.push({
+            point: { x, y },
+            label: { text, width, height },
+        });
+    }
+
+    return labelData;
 }
